+++ CHANGELOG.rst
@@ -3,10 +3,11 @@ Changelog
 
 Here you can see the full list of changes between each aerofiles release.
 
-aerofiles v1.5.X, 2025-07-XX
+aerofiles v1.5.0, 2026-03-XX
 ----------------------------
 * AIXM airspace reader.
 * "make lint-fix" for autopep8 and applied to all sources
+* add "AA" (activation times) handling to openair
 
 aerofiles v1.4.3, 2025-06-23
 ----------------------------